Separates out pointer-generator modules.
There is no RNN pointer generator decoder module, but there is a
separate one for transformers (and a rotary variant). We put this in a
separate file to reduce the considerable length of the transformer
module file by about 200 lines.

Some minor drivebys are applied.
