+++ yoyodyne/models/modules/hard_attention.py
@@ -295,7 +295,9 @@ class HardAttentionTransformerDecoder(transformer.TransformerModule):
     output_proj: nn.Sequential
     scale_encoded: nn.Linear
 
-    def __init__(self, *args, decoder_input_size: int, **kwargs):
+    def __init__(
+        self, *args, decoder_input_size: int = defaults.HIDDEN_SIZE, **kwargs
+    ):
         # Must be set before super().__init__ so output_size is available
         # when output_proj and scale_encoded are built below.
         self.decoder_input_size = decoder_input_size
